introduce proxy to delegate endpoint implementation
