+++ postcard_creator/postcard_creator_swissid.py
@@ -1,0 +1,37 @@
+import requests
+
+from postcard_creator.postcard_creator import PostcardCreatorException, _dump_request, logger, PostcardCreatorImpl
+
+
+class PostcardCreatorSwissId(PostcardCreatorImpl):
+    def __init__(self, token=None):
+        if token.token is None:
+            raise PostcardCreatorException('No Token given')
+        self.token = token
+        self._session = self._create_session()
+
+    def _get_headers(self):
+        return {
+            'User-Agent': 'Mozilla/5.0 (Linux; Android 6.0.1; wv) AppleWebKit/537.36 (KHTML, like Gecko) '
+                          'Version/4.0 Chrome/52.0.2743.98 Mobile Safari/537.36',
+            'Authorization': 'Bearer {}'.format(self.token.token)
+        }
+
+    def _create_session(self):
+        return requests.Session()
+
+    def _do_op(self, method, endpoint, **kwargs):
+        url = self.host + endpoint
+        if 'headers' not in kwargs or kwargs['headers'] is None:
+            kwargs['headers'] = self._get_headers()
+
+        logger.debug('{}: {}'.format(method, url))
+        response = self._session.request(method, url, **kwargs)
+        _dump_request(response)
+
+        if response.status_code not in [200, 201, 204]:
+            e = PostcardCreatorException('error in request {} {}. status_code: {}'
+                                         .format(method, url, response.status_code))
+            e.server_response = response.text
+            raise e
+        return response
